feat: RacingGameController#run 구현
Application#main에 구현된 메서드를 Controller 클래스로 이동
